rapport à jour

+++ src/main/java/com/poste/ProjetIPM/entities/IPM_StatutEmploye.java
@@ -15,7 +15,7 @@ import java.util.List;
 public class IPM_StatutEmploye {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idStatut;
+    private Long idStatut;
     private String emplStatut;
     @JsonIgnore
     @OneToMany(mappedBy = "ipmStatutEmploye")
